moved back to pull request and added integration test (#122)
* moved back to pull request and added integration test


* Integration test seperation (#123)

* decouple resource and resource inner  (#86)

* decoupled resources and watch


* seperation of integration and unit tests
